agent_v5/requirements.txt
langchain==1.2.8
langchain-core==1.2.8
langchain-community==0.4.1
langchain-ollama==1.0.1
pydantic>=2.7.0,<3
datasets>=2.19.0
tqdm>=4.66.0
sympy>=1.12
python-dotenv>=1.0.0
ipykernel>=6.29.0
matplotlib>=3.8.0
nvidia-ml-py>=12.0.0
agentlightning>=0.3.0
openai>=1.0.0
